Fixes Publicacion
- Mapper: borrado de metodos viejos
- eliminacion de cosas del Repository que no se podian resolver

+++ src/main/java/is2/g57/hopetrade/repository/PublicacionStateRepository.java
@@ -4,6 +4,6 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import is2.g57.hopetrade.entity.state.PublicacionState;
 
-public interface PublicacionStateRepository extends JpaRepository<PublicacionState, Integer> {
+public interface PublicacionStateRepository extends JpaRepository<PublicacionState, Long> {
     Optional<PublicacionState> findByNombre(String name);
 }
